requirements.txt
asgiref==3.7.2
cairocffi==1.6.1
CairoSVG==2.7.1
cffi==1.16.0
chardet==5.2.0
crispy-bootstrap4==2023.1
cssselect2==0.7.0        
defusedxml==0.7.1        
distlib==0.3.7
Django==4.2.7
django-chartjs==2.3.0
django-crispy-forms==2.1
django-phonenumber-field==7.3.0
filelock==3.13.1
importlib-metadata==7.0.1
mysqlclient==2.2.0
phonenumbers==8.13.27
Pillow==10.1.0
platformdirs==4.0.0
pycparser==2.21
pygal==3.0.4
reportlab==4.0.8
sqlparse==0.4.4
tinycss2==1.2.1
typing-extensions==4.8.0
tzdata==2023.3
virtualenv==20.24.7
webencodings==0.5.1
zipp==3.17.0